Added sliderHeight public variable to ImageWindow
ImageCanvas, ImageWindow, and Zoom would guess whether the window had
scrollbars at the bottom, and assign an extra 20 pixels if it was.
However, hyperstacks can have more scrollbars. Therefore I added a
public variable sliderHeight to ImageWindow, which should keep track of
the total height of the scrollbars plus gaps in between. The StackWindow
class updates sliderHeight by adding the height of the scrollbar plus
the gap. I also added a public final static HGAP and VGAP to ImageWindow
so they would be easy to change if needed. This in turn required a small
change to ImageLayout.